Add files via upload
to make it work, first use your openAI API key and put it in the .env file. then download appropriate pdfs of cars (only 4-5) and declare the path on the 294th line. run "uv run main.py" in the terminal and wait for the langchain and AI to read and make packets, then ask appropriate queries!